+++ src/amun/common/sentence.h
@@ -11,15 +11,15 @@ class God;
 class Sentence {
   public:
 
-    Sentence(const God &god, size_t vLineNum, const std::string& line);
-    Sentence(const God &god, size_t vLineNum, const std::vector<std::string>& words);
-		Sentence(God &god, size_t lineNum, const std::vector<uint>& words);
+    Sentence(const God &god, unsigned vLineNum, const std::string& line);
+    Sentence(const God &god, unsigned vLineNum, const std::vector<std::string>& words);
+		Sentence(God &god, unsigned lineNum, const std::vector<uint>& words);
 
-    const Words& GetWords(size_t index = 0) const;
-    const FactWords& GetFactors(size_t index = 0) const;
-    size_t size(size_t index = 0) const;
+    const Words& GetWords(unsigned index = 0) const;
+    const FactWords& GetFactors(unsigned index = 0) const;
+    unsigned size(unsigned index = 0) const;
 
-    size_t GetLineNum() const;
+    unsigned GetLineNum() const;
 
 
   private:
@@ -27,7 +27,7 @@ class Sentence {
 
     std::vector<Words> words_;
     std::vector<FactWords> factors_;
-    size_t lineNum_;
+    unsigned lineNum_;
 
     Sentence(const Sentence &) = delete;
 };
